refactor: 이름 관련 정책 구현 위치 view로 이동
