+++ README.md
@@ -66,5 +66,5 @@
 - [ ] 도메인 로직에서 UI 로직 분리
     - [ ] ~ String 클래스 이름 변경
     - [ ] ~ String 클래스 view 패키지로 이동
-    - [ ] ~ 이름 관련 정책 구현 위치 view로 이동
+    - [x] 이름 관련 정책 구현 위치 view로 이동
 - [ ] 누락된 테스트 추가
