+++ src/main/java/Main.java
@@ -1,16 +1,18 @@
 import domain.LadderGame;
+import java.util.List;
 import java.util.function.Supplier;
 import view.InputView;
+import view.NameInputView;
 import view.OutPutView;
 
 public class Main {
     public static void main(String[] args) {
         LadderGame ladderGame = RetryHelper.retry(() -> {
             OutPutView.print("참여할 사람 이름을 입력하세요. (이름은 쉼표(,)로 구분하세요)");
-            String nameInput = InputView.getInput();
+            List<String> names = NameInputView.getNames(InputView::getInput);
             OutPutView.print("최대 사다리 높이는 몇 개인가요?");
             int ladderHeight = Integer.parseInt(InputView.getInput());
-            return new LadderGame(nameInput, ladderHeight);
+            return new LadderGame(names, ladderHeight);
         });
         OutPutView.print(ladderGame.getNamesString());
         OutPutView.print(ladderGame.getLadderString());
