+++ src/main/java/domain/LadderGame.java
@@ -1,10 +1,12 @@
 package domain;
 
+import java.util.List;
+
 public class LadderGame {
     private final Ladder ladder;
     private final Names names;
 
-    public LadderGame(String userNames, int ladderHeight) {
+    public LadderGame(List<String> userNames, int ladderHeight) {
         names = new Names(userNames);
         int nameCount = names.getNameCount();
         Height height = new Height(ladderHeight);
