+++ src/test/java/domain/NamesStringTest.java
@@ -1,5 +1,6 @@
 package domain;
 
+import java.util.List;
 import org.assertj.core.api.Assertions;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
@@ -8,7 +9,7 @@ class NamesStringTest {
     @Test
     @DisplayName("이름 문자열 생성")
     void test() {
-        Names names = new Names("a,aa,aaa,aaaa,aaaaa");
+        Names names = new Names(List.of("a", "aa", "aaa", "aaaa", "aaaaa"));
         String actual = NamesString.from(names);
         String expected = "   a    aa   aaa  aaaa  aaaaa";
         Assertions.assertThat(actual).isEqualTo(expected);
